Updated Repulsionbot; improved encrypt function and added bug functionality.
Encrypt function now no longer requires quotes around the sent message text, and also checks for any bugs placed on the sender or receiver. 

Added bug functionality integrated with nicknames database:
'Bugs' can now be placed on another player, with the caller of the function being designated the setter, and by default give 3 uses. Bugs will send encrypted messages that are sent or received by the person the bug was placed on to the setter.
Ref functionality added to the bugs, with the ability to: 
- View all listed bugs.
- Place bugs for another person with a designated number of uses.
- Delete specific bugs.
- Purge all bugs.
If someone who is not a ref tries to access these functions, it will not allow it, and can send a message to a designated channel with a note of who attempted to call the function.